requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2
